requirements.txt
streamlit==1.34.0
python-dotenv==1.0.0
pandas==2.2.2
openpyxl==3.1.2
google-generativeai==0.14.1
